assets/style.css: make the text easier for the eyes
Larger blocks of white text on black background were hard to read. Reduce
the contrast by making the text grey for the body. Headers and links are
left as they were.

This commit also cleans up the CSS a bit by moving variables to global
scope and unifying spelling of grey/gray.

Signed-off-by: Krystian Hebel <krystian.hebel@3mdeb.com>